Extract tests for trim functions into a separate file
